requirements.txt
asgiref==3.8.1
Django==5.2
django-currentuser==0.9.0
django-environ==0.12.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
gunicorn==23.0.0
packaging==24.2
psycopg2-binary==2.9.10
PyJWT==2.9.0
sqlparse==0.5.3
whitenoise==6.9.0
